feat: extract DIalog Content to be usable by other flows
