+++ .changeset/sweet-years-camp.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": minor
+---
+
+Wrap ModularAsset content to be usable by other flow like Send
